backend/chat-app/requirements.txt
python~=3.11.11
quart~=0.20.0
asyncpg~=0.30.0
python-dotenv~=1.0.1